lp-lxc - added much-needed restart safeguard, `lxc_restart`
lxc_restart can be filled with a prompt and need to be explicitly
set to 'YES' (in uppercase!) to make the role automatically
configure container networking after their initial creation. This
can be offered to the user in a prompt, an example of which was
added to lp-lxc/README.md.